chore(project): rename project from dona-matic to dona-pol
This commit updates the project's internal name in the `package.json` and other relevant configuration files to 'dona-pol'.

The change was made to align the project's identity with the evolution of the Polygon network, where the native token has transitioned from MATIC to POL. This new name better reflects the platform on which the DApp operates.